commit project improvements

+++ Store.Assignment/Products/Beverage.cs
@@ -2,7 +2,7 @@
 {
     public class Beverage : PerishableProduct
     {
-        public Beverage(string name, string brand, double price, int quantity, DateTime expirationDate) 
+        public Beverage(string name, string brand, double price, double quantity, DateTime expirationDate) 
             : base(name, brand, price, quantity, expirationDate)
         {
         }
