+++ Store.Assignment/Products/Food.cs
@@ -2,7 +2,7 @@
 {
     public class Food : PerishableProduct
     {
-        public Food(string name, string brand, double price, int quantity, DateTime expirationDate) 
+        public Food(string name, string brand, double price, double quantity, DateTime expirationDate) 
             : base(name, brand, price, quantity, expirationDate)
         {
         }
